applying review updates
